fix(react): Use correct router
Using the `BrowserRouter` will cause errors because it create it's own
history object and sets that globally. This means it will interfere with
the history created by `createBrowserHistory()`, which is what Sentry
uses to navigate. To fix this, we switch to using the base `Router`,
which doesn't create it's own history, instead relies on the
external history we provide it.